+++ packages/component-library/src/components/form/mt-text-field/mt-text-field.vue
@@ -260,9 +260,8 @@ export default defineComponent({
       this.$emit("update:modelValue", event.target.value);
     },
 
-    handleFocus(event: Event): void {
-      this.$emit("focus", event);
-      this.setFocusClass();
+    handleFocus(event: FocusEvent): void {
+      this.setFocusClass(event);
     },
 
     restoreInheritance(): void {
@@ -277,11 +276,13 @@ export default defineComponent({
       return `${identification}-${this.idSuffix}`;
     },
 
-    setFocusClass(): void {
+    setFocusClass(event: FocusEvent): void {
+      this.$emit("focus", event);
       this.hasFocus = true;
     },
 
-    removeFocusClass(): void {
+    removeFocusClass(event: FocusEvent): void {
+      this.$emit("blur", event);
       this.hasFocus = false;
     },
   },