+++ packages/icon-kit/src/filesystem/in-memory-filessytem.ts
@@ -1,13 +1,40 @@
 import type { Filesystem } from "./filesystem.js";
 
+type Entry =
+  | {
+      type: "file";
+      content: string;
+    }
+  | {
+      type: "directory";
+    };
+
 export class InMemoryFilesystem implements Filesystem {
-  private files: Record<string, string> = {};
+  private files: Record<string, Entry> = {};
 
   createFile(path: string, content: string) {
-    this.files[path] = content;
+    this.files[path] = {
+      type: "file",
+      content,
+    };
   }
 
-  readFile(path: string) {
-    return this.files[path];
+  readFile(path: string): string | undefined {
+    const entry = this.files[path];
+    if (entry?.type === "file") {
+      return entry.content;
+    }
+
+    return undefined;
+  }
+
+  createDirectory(path: string) {
+    this.files[path] = {
+      type: "directory",
+    };
+  }
+
+  removeDirectory(path: string) {
+    delete this.files[path];
   }
 }